API/requirements.txt
Flask
selenium
beautifulsoup4
spacy
python-dotenv